Fixed a bug when reading the GMFs from a file: the hazard sites were not used
